# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Amazonet",
        "description": "This platform, similar to Amazon. It enables user authentication, product management, catalog browsing, shopping cart and checkout processes, order management, and more. The platform supports multiple roles, including customers, deliveryman, and administrators, providing each with specific features to manage products, orders, and user interactions."
    },
    "servers": [
        {
            "url": "http://localhost:4000/"
        }
    ],
...
    "paths": {
...
        "/products/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get product whose identifier is known",
                "description": "Get product whose identifier is known",
                "operationId": "getProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "product": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found!"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "products"
                ],
                "summary": "Update product whose identifier is known",
                "description": "Update product whose identifier is known",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The product that needs to updated",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Product with new properties",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "example": "image_product"
                                    },
                                    "countInStock": {
                                        "type": "number",
                                        "example": 10
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "description_product"
                                    },
                                    "discount": {
                                        "type": "floot",
                                        "example": 12.5
                                    },
                                    "priceDiscount": {
                                        "type": "floot",
                                        "example": 49.000
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product updated"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bad request!"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found!"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "products"
                ],
                "summary": "Delete product whose identifier is known",
                "description": "Delete product whose identifier is known",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The product that needs to be deleted",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product is deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product is deleted"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product not found!"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/products/filter/search": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "products"
                ],
                "summary": "Filter and search for products",
                "description": "Filter products based on category, price, rating, and search query. Also supports sorting and pagination.",
                "operationId": "filterProduct",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "description": "Page number",
                        "type": "integer"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "type": "integer",
                        "description": "Number of items per page"
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "description": "Category of product"
                    },
                    {
                        "name": "minPrice",
                        "in": "query",
                        "description": "Filter users created great then this price",
                        "required": false,
                        "type": "floot"
                    },
                    {
                        "name": "maxPrice",
                        "in": "query",
                        "description": "Filter users created less then this price",
                        "required": false,
                        "type": "floot"
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "required": false,
                        "type": "string",
                        "description": "Rating of the product"
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "featured",
                                "lowest",
                                "highest",
                                "toprated",
                                "newest"
                            ],
                            "default": "featured"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "listProducts": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "products": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid price range!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid price range!"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "There are no product!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "There are no product!"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "required": [
                    "name",
                    "slug",
                    "brand",
                    "category",
                    "description",
                    "price",
                    "countInStock",
                    "rating",
                    "numReviews"
                ],
                "type": "object",
                "properties": {
                    "_id": {
                        "type": "string",
                        "example": "66c1c679b966da0f213a2aad"
                    },
                    "name": {
                        "type": "string",
                        "example": "customer"
                    },
                    "slug": {
                        "type": "string",
                        "example": "slug"
                    },
                    "image": {
                        "type": "string",
                        "example": "Product Image"
                    },
                    "category": {
                        "type": "string",
                        "example": "Electronics",
                        "enum": [
                            "Electronics",
                            "Computers",
                            "Women's Fashion",
                            "Men's Fashion",
                            "Video Games",
                            "Books"
                        ]
                    },
                    "description": {
                        "type": "string",
                        "example": "description"
                    },
                    "price": {
                        "type": "integer",
                        "example": "20.000"
                    },
                    "discount": {
                        "type": "integer",
                        "example": "10.5"
                    },
                    "priceDiscount": {
                        "type": "integer",
                        "example": "4.50"
                    },
                    "countInStock": {
                        "type": "number",
                        "example": "3"
                    },
                    "rating": {
                        "type": "number",
                        "example": "3"
                    },
                    "numReviews": {
                        "type": "number",
                        "example": "3"
                    },
                    "reviews": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Review"
                        }
                    }
                }
            },
            "Review": {
                "required": [
                    "name",
                    "comment",
                    "rating"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "user_name"
                    },
                    "comment": {
                        "type": "string",
                        "example": "comment of product"
                    },
                    "rating": {
                        "type": "number",
                        "example": 3
                    }
                }
            },
...
        },
...
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "authorization",
                "in": "header"
            }
        }
    }
}